pyproject.toml
[tool.poetry]
name = "ticketease"
version = "0.1.0"
description = ""
authors = ["rohanathan.s@hotmail.com <rohanthan.s@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
psycopg2-binary = "^2.9.10"
django-environ = "^0.12.0"
pillow = "^11.1.0"
qrcode = {extras = ["pil"], version = "^8.0"}
djangorestframework = "^3.15.2"
stripe = "^11.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
